pyproject.toml
[project]
name = "personal-space-api"
version = "0.1.0"
description = "Backend API for Personal Space dashboard"
requires-python = ">=3.12"
dependencies = [
    # Framework
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Auth
    "firebase-admin>=6.6.0",
    # Security
    "bcrypt>=4.2.0",
    "cryptography>=44.0.0",
    "pyjwt>=2.9.0",
    # Utilities
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.18",
    # Rate limiting
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
